Refactor models page
